refactor: ProductAnalysis 관련 DTO 이동
