+++ src/main/java/com/webeye/backend/cosmetic/application/CosmeticService.java
@@ -1,7 +1,7 @@
 package com.webeye.backend.cosmetic.application;
 
 import com.webeye.backend.cosmetic.dto.response.CosmeticResponse;
-import com.webeye.backend.explanation.dto.request.ProductAnalysisRequest;
+import com.webeye.backend.product.dto.request.ProductAnalysisRequest;
 
 public interface CosmeticService {
     CosmeticResponse analyzeCosmetic(ProductAnalysisRequest request);
