+++ src/main/java/com/webeye/backend/cosmetic/application/CosmeticServiceImpl.java
@@ -1,8 +1,8 @@
 package com.webeye.backend.cosmetic.application;
 
 import com.webeye.backend.cosmetic.dto.response.CosmeticResponse;
-import com.webeye.backend.explanation.dto.request.ProductAnalysisRequest;
 import com.webeye.backend.imageanalysis.infrastructure.OpenAiClient;
+import com.webeye.backend.product.dto.request.ProductAnalysisRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
