+++ src/main/java/com/webeye/backend/cosmetic/presentation/CosmeticController.java
@@ -3,8 +3,8 @@ package com.webeye.backend.cosmetic.presentation;
 import com.webeye.backend.cosmetic.application.CosmeticService;
 import com.webeye.backend.cosmetic.dto.response.CosmeticResponse;
 import com.webeye.backend.cosmetic.presentation.swagger.CosmeticSwagger;
-import com.webeye.backend.explanation.dto.request.ProductAnalysisRequest;
 import com.webeye.backend.global.dto.response.SuccessResponse;
+import com.webeye.backend.product.dto.request.ProductAnalysisRequest;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
