+++ src/main/java/com/webeye/backend/product/dto/request/ProductDetailAnalysisRequest.java
@@ -1,4 +1,4 @@
-package com.webeye.backend.explanation.dto.request;
+package com.webeye.backend.product.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
@@ -6,14 +6,11 @@ import lombok.Builder;
 
 import java.util.List;
 
-@Schema(description = "제품 분석")
+@Schema(description = "상품 설명 이미지의 URL")
 @Builder
-public record ProductAnalysisRequest(
-        @NotEmpty(message = "제품 ID는 비어있을 수 없습니다.")
-        String productId,
-
+public record ProductDetailAnalysisRequest(
         @Schema(description = "상품 이미지 URL")
         @NotEmpty(message = "이미지 URL 목록은 비어있을 수 없습니다.")
         List<String> urls
-) {
+){
 }
